requirements/prod.txt
# Everything needed in production

# Flask
Flask==1.1.2
Werkzeug==1.0.1
click>=7.0

# Database
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.19
psycopg2-binary==2.8.6

# Migrations
Flask-Migrate==2.5.3

# Forms
email-validator==1.1.1
Flask-WTF==0.14.3
WTForms==2.3.3

# Deployment
gevent==20.9.0
gunicorn>=19.9.0
supervisor==4.2.1

# Flask Static Digest
Flask-Static-Digest==0.1.3

# Auth
Flask-Login==0.5.0
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.7.2

# Debug toolbar
Flask-DebugToolbar==0.11.0

# Environment variable parsing
environs==8.0.0
PyYAML==5.3.1
WTForms-Alchemy==0.17.0
Flask-Uploads@git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc93e684fa0a3a4350a38e41ae00483a796
awesome-slugify==1.6.5
mysql-connector==2.2.9
Flask-RESTful==0.3.8
Flask-Session@git+https://github.com/rayluo/flask-session@ea8351438e4c79bad59429a43dd662c167c32a32
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.0
GitPython==3.1.11
flask-reverse-proxy-fix@git+https://github.com/eibrahim995/flask-reverse-proxy-fix.git@8dafa59f34600b2dfd5ef5067997677ffd1d58b6
flask-rest-paginate==0.1.5
pyopenssl
redis==3.5.3
rq==1.6.1
Flask-RQ==0.2
python-ffmpeg-video-streaming==0.1.14
ffmpeg-python===0.2.0
Flask-SocketIO==4.3.1
python-socketio==4.6.0
gevent-websocket==0.10.1
rq-scheduler
requests
Flask-RQ2
